Remove coredata code generation
